fix(go-core): create llm resources via gateway
