+++ suites/go-core/tests/gateway_helpers_test.go
@@ -1,4 +1,4 @@
-//go:build e2e && (svc_gateway || svc_llm || svc_llm_proxy || smoke)
+//go:build e2e && (svc_gateway || svc_agents_orchestrator || svc_runners || svc_metering || svc_k8s_runner || svc_organizations || svc_llm || svc_llm_proxy || smoke)
 
 package tests
